Move cross-circuit table stuff to src/table.rs
