+++ zkevm-circuits/src/evm_circuit.rs
@@ -11,20 +11,11 @@ pub(crate) mod util;
 pub mod table;
 pub mod witness;
 
-use crate::bytecode_circuit::bytecode_unroller::BytecodeTable;
-use crate::tx_circuit::TxTable;
-use crate::{
-    evm_circuit::{
-        util::{rlc, RandomLinearCombination},
-        witness::{BlockContext, Bytecode, RwMap, Transaction},
-    },
-    rw_table::RwTable,
-};
-use eth_types::{Field, ToLittleEndian, Word};
+use crate::table::{BytecodeTable, TxTable};
+use eth_types::Field;
 use execution::ExecutionConfig;
 use itertools::Itertools;
-use keccak256::plain::Keccak;
-use table::{BlockTable, FixedTableTag, KeccakTable, LookupTable, TableColumns};
+use table::{FixedTableTag, LookupTable};
 use witness::Block;
 
 /// EvmCircuit implements verification of execution trace of a block.
@@ -149,244 +140,12 @@ impl<F: Field> EvmCircuit<F> {
     }
 }
 
-// TODO: Move to src/tables.rs
-pub fn load_txs<F: Field>(
-    tx_table: &TxTable,
-    layouter: &mut impl Layouter<F>,
-    txs: &[Transaction],
-    randomness: F,
-) -> Result<(), Error> {
-    layouter.assign_region(
-        || "tx table",
-        |mut region| {
-            let mut offset = 0;
-            for column in tx_table.columns() {
-                region.assign_advice(
-                    || "tx table all-zero row",
-                    column,
-                    offset,
-                    || Ok(F::zero()),
-                )?;
-            }
-            offset += 1;
-
-            // println!("DBG load_txs");
-            let tx_table_columns = tx_table.columns();
-            for tx in txs.iter() {
-                for row in tx.table_assignments(randomness) {
-                    // print!("{:02} ", offset);
-                    for (column, value) in tx_table_columns.iter().zip_eq(row) {
-                        // print!("{:?} ", value);
-                        region.assign_advice(
-                            || format!("tx table row {}", offset),
-                            *column,
-                            offset,
-                            || Ok(value),
-                        )?;
-                    }
-                    offset += 1;
-                    // println!("");
-                }
-            }
-            Ok(())
-        },
-    )
-}
-
-// TODO: Move to src/tables.rs
-pub fn load_rws<F: Field>(
-    rw_table: &RwTable,
-    layouter: &mut impl Layouter<F>,
-    rws: &RwMap,
-    randomness: F,
-) -> Result<(), Error> {
-    layouter.assign_region(
-        || "rw table",
-        |mut region| {
-            let mut offset = 0;
-            rw_table.assign(&mut region, offset, &Default::default())?;
-            offset += 1;
-
-            let mut rows = rws
-                .0
-                .values()
-                .flat_map(|rws| rws.iter())
-                .collect::<Vec<_>>();
-
-            rows.sort_by_key(|a| a.rw_counter());
-            let mut expected_rw_counter = 1;
-            for rw in rows {
-                assert!(rw.rw_counter() == expected_rw_counter);
-                expected_rw_counter += 1;
-
-                rw_table.assign(&mut region, offset, &rw.table_assignment(randomness))?;
-                offset += 1;
-            }
-            Ok(())
-        },
-    )
-}
-
-// use crate::util::TableShow;
-
-// TODO: Move to src/tables.rs
-pub fn load_bytecodes<'a, F: Field>(
-    bytecode_table: &BytecodeTable,
-    layouter: &mut impl Layouter<F>,
-    bytecodes: impl IntoIterator<Item = &'a Bytecode> + Clone,
-    randomness: F,
-) -> Result<(), Error> {
-    // println!("> load_bytecodes");
-    // let mut table = TableShow::<F>::new(vec!["codeHash", "tag", "index",
-    // "isCode", "value"]);
-    layouter.assign_region(
-        || "bytecode table",
-        |mut region| {
-            let mut offset = 0;
-            for column in bytecode_table.columns() {
-                region.assign_advice(
-                    || "bytecode table all-zero row",
-                    column,
-                    offset,
-                    || Ok(F::zero()),
-                )?;
-            }
-            offset += 1;
-
-            let bytecode_table_columns = bytecode_table.columns();
-            for bytecode in bytecodes.clone() {
-                for row in bytecode.table_assignments(randomness) {
-                    // let mut column_index = 0;
-                    for (column, value) in bytecode_table_columns.iter().zip_eq(row) {
-                        region.assign_advice(
-                            || format!("bytecode table row {}", offset),
-                            *column,
-                            offset,
-                            || Ok(value),
-                        )?;
-                        // table.push(column_index, value);
-                        // column_index += 1;
-                    }
-                    offset += 1;
-                }
-            }
-            // table.print();
-            Ok(())
-        },
-    )
-}
-
-// TODO: Move to src/tables.rs
-pub fn load_block<F: Field>(
-    block_table: &BlockTable,
-    layouter: &mut impl Layouter<F>,
-    block: &BlockContext,
-    randomness: F,
-) -> Result<(), Error> {
-    layouter.assign_region(
-        || "block table",
-        |mut region| {
-            let mut offset = 0;
-            for column in block_table.columns() {
-                region.assign_advice(
-                    || "block table all-zero row",
-                    column,
-                    offset,
-                    || Ok(F::zero()),
-                )?;
-            }
-            offset += 1;
-
-            let block_table_columns = block_table.columns();
-            for row in block.table_assignments(randomness) {
-                for (column, value) in block_table_columns.iter().zip_eq(row) {
-                    region.assign_advice(
-                        || format!("block table row {}", offset),
-                        *column,
-                        offset,
-                        || Ok(value),
-                    )?;
-                }
-                offset += 1;
-            }
-
-            Ok(())
-        },
-    )
-}
-
-pub fn keccak_table_assignments<F: Field>(input: &[u8], randomness: F) -> Vec<[F; 4]> {
-    // CHANGELOG: Using `RLC(reversed(input))`
-    let input_rlc: F = rlc::value(input.iter().rev(), randomness);
-    let input_len = F::from(input.len() as u64);
-    let mut keccak = Keccak::default();
-    keccak.update(input);
-    let output = keccak.digest();
-    let output_rlc = RandomLinearCombination::<F, 32>::random_linear_combine(
-        Word::from_big_endian(output.as_slice()).to_le_bytes(),
-        randomness,
-    );
-
-    vec![[F::one(), input_rlc, input_len, output_rlc]]
-}
-
-// NOTE: For now, the input_rlc of the keccak is defined as
-// `RLC(reversed(input))` for convenience of the circuits that do the lookups.
-// This allows calculating the `input_rlc` after all the inputs bytes have been
-// layed out via the pattern `acc[i] = acc[i-1] * r + value[i]`.
-pub fn load_keccaks<'a, F: Field>(
-    keccak_table: &KeccakTable,
-    layouter: &mut impl Layouter<F>,
-    inputs: impl IntoIterator<Item = &'a [u8]> + Clone,
-    randomness: F,
-) -> Result<(), Error> {
-    // println!("> super_circuit.load_keccaks");
-    // let mut table = TableShow::<F>::new(vec!["is_enabled", "input_rlc",
-    // "input_len", "output_rlc"]);
-    layouter.assign_region(
-        || "keccak table",
-        |mut region| {
-            let mut offset = 0;
-            for column in keccak_table.columns() {
-                region.assign_advice(
-                    || "keccak table all-zero row",
-                    column,
-                    offset,
-                    || Ok(F::zero()),
-                )?;
-            }
-            offset += 1;
-
-            let keccak_table_columns = keccak_table.columns();
-            for input in inputs.clone() {
-                // println!("+ {:?}", input);
-                for row in keccak_table_assignments(input, randomness) {
-                    // let mut column_index = 0;
-                    for (column, value) in keccak_table_columns.iter().zip_eq(row) {
-                        region.assign_advice(
-                            || format!("keccak table row {}", offset),
-                            *column,
-                            offset,
-                            || Ok(value),
-                        )?;
-                        // table.push(column_index, value);
-                        // column_index += 1;
-                    }
-                    offset += 1;
-                }
-            }
-            // table.print();
-            Ok(())
-        },
-    )
-}
-
 #[cfg(any(feature = "test", test))]
 pub mod test {
     use super::*;
     use crate::{
         evm_circuit::{table::FixedTableTag, witness::Block, EvmCircuit},
-        rw_table::RwTable,
+        table::{load_block, load_bytecodes, load_rws, load_txs, BlockTable, RwTable},
         util::power_of_randomness_from_instance,
     };
     use eth_types::{Field, Word};
