+++ zkevm-circuits/src/evm_circuit/execution/calldatacopy.rs
@@ -3,7 +3,6 @@ use crate::{
         execution::ExecutionGadget,
         param::{N_BYTES_MEMORY_ADDRESS, N_BYTES_MEMORY_WORD_SIZE},
         step::ExecutionState,
-        table::CallContextFieldTag,
         util::{
             common_gadget::SameContextGadget,
             constraint_builder::{
@@ -16,6 +15,7 @@ use crate::{
         },
         witness::{Block, Call, ExecStep, Transaction},
     },
+        table::CallContextFieldTag,
     util::Expr,
 };
 use bus_mapping::evm::OpcodeId;
