+++ zkevm-circuits/src/evm_circuit/execution/calldatasize.rs
@@ -1,9 +1,9 @@
 use crate::{
+        table::CallContextFieldTag,
     evm_circuit::{
         execution::ExecutionGadget,
         param::N_BYTES_CALLDATASIZE,
         step::ExecutionState,
-        table::CallContextFieldTag,
         util::{
             common_gadget::SameContextGadget,
             constraint_builder::{ConstraintBuilder, StepStateTransition, Transition::Delta},
