+++ zkevm-circuits/src/evm_circuit/execution/copy_to_log.rs
@@ -3,7 +3,6 @@ use crate::{
         execution::ExecutionGadget,
         param::{N_BYTES_MEMORY_ADDRESS, N_BYTES_MEMORY_WORD_SIZE},
         step::ExecutionState,
-        table::TxLogFieldTag,
         util::{
             constraint_builder::{ConstraintBuilder, StepStateTransition, Transition::Delta},
             math_gadget::ComparisonGadget,
@@ -12,6 +11,7 @@ use crate::{
         },
         witness::{Block, Call, ExecStep, Transaction},
     },
+        table::TxLogFieldTag,
     util::Expr,
 };
 use bus_mapping::{circuit_input_builder::CopyDetails, constants::MAX_COPY_BYTES};
@@ -229,12 +229,13 @@ impl<F: Field> ExecutionGadget<F> for CopyToLogGadget<F> {
 
 #[cfg(test)]
 pub mod test {
-    use crate::evm_circuit::{
+    use crate::{evm_circuit::{
         step::ExecutionState,
-        table::{RwTableTag, TxLogFieldTag},
         test::{rand_bytes, run_test_circuit_incomplete_fixed_table},
         witness::{Block, Bytecode, Call, ExecStep, Rw, RwMap, Transaction},
-    };
+    },
+        table::{RwTableTag, TxLogFieldTag},
+};
     use bus_mapping::{
         circuit_input_builder::{CopyDetails, StepAuxiliaryData},
         constants::MAX_COPY_BYTES,
