+++ zkevm-circuits/src/evm_circuit/execution/memory_copy.rs
@@ -1,9 +1,9 @@
 use crate::{
+        table::TxContextFieldTag,
     evm_circuit::{
         execution::ExecutionGadget,
         param::{N_BYTES_MEMORY_ADDRESS, N_BYTES_MEMORY_WORD_SIZE},
         step::ExecutionState,
-        table::TxContextFieldTag,
         util::{
             constraint_builder::{ConstraintBuilder, StepStateTransition, Transition::Delta},
             math_gadget::ComparisonGadget,
@@ -235,13 +235,14 @@ impl<F: Field> ExecutionGadget<F> for CopyToMemoryGadget<F> {
 
 #[cfg(test)]
 pub mod test {
-    use crate::evm_circuit::{
+    use crate::{evm_circuit::{
         execution::memory_copy::MAX_COPY_BYTES,
         step::ExecutionState,
-        table::RwTableTag,
         test::{rand_bytes, run_test_circuit_incomplete_fixed_table},
         witness::{Block, Bytecode, Call, ExecStep, Rw, RwMap, Transaction},
-    };
+    },
+    table::RwTableTag,
+};
     use bus_mapping::circuit_input_builder::{CopyDetails, StepAuxiliaryData};
     use eth_types::evm_types::OpcodeId;
     use halo2_proofs::arithmetic::BaseExt;
