+++ zkevm-circuits/src/evm_circuit/execution/sload.rs
@@ -1,8 +1,8 @@
 use crate::{
+        table::CallContextFieldTag,
     evm_circuit::{
         execution::ExecutionGadget,
         step::ExecutionState,
-        table::CallContextFieldTag,
         util::{
             common_gadget::SameContextGadget,
             constraint_builder::{
