+++ zkevm-circuits/src/lib.rs
@@ -21,6 +21,7 @@ pub mod evm_circuit;
 pub mod rw_table;
 pub mod state_circuit;
 pub mod super_circuit;
+pub mod table;
 #[cfg(test)]
 pub mod test_util;
 pub mod tx_circuit;
