+++ zkevm-circuits/src/state_circuit/constraint_builder.rs
@@ -3,10 +3,11 @@ use super::{
     random_linear_combination::Queries as RlcQueries, N_LIMBS_ACCOUNT_ADDRESS, N_LIMBS_ID,
     N_LIMBS_RW_COUNTER,
 };
-use crate::evm_circuit::{
+use crate::{evm_circuit::{
     param::N_BYTES_WORD,
-    table::{AccountFieldTag, RwTableTag},
     util::not,
+},
+    table::{AccountFieldTag, RwTableTag},
 };
 use crate::util::Expr;
 use eth_types::Field;
